requirements.txt
Django==1.8.2
Pillow==2.8.2
argparse==1.2.1
django-gcm==1.0.7
django-tastypie==0.12.2
djangorestframework==3.1.3
feedparser==5.2.1
gunicorn==19.3.0
psycopg2==2.6.1
python-crontab==1.9.3
python-dateutil==2.4.2
python-ldap==2.4.20
python-mimeparse==0.1.4
pytz==2015.4
requests==2.7.0
six==1.9.0
wheel==0.24.0
wsgiref==0.1.2
